feat(Order Products CMP)
Added  OrderItemAddedEvent event listener and handler.

It seems that the apex.refresh, and even the re-invoke of the Apex method
to update the datatable info doesn't work if the initial invoke is done via
@wire decorator. To solve this, the apex method to retrieve de data of the
order item is invoke imperatively.

As result, remove the (cacheable=true) param of the AuraEnabled annotation
is needed.

Finally, to recharge the data in tne table, a spinner is added in the
html file, to show it while the table is rebuilted.